requirements.txt
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
flasgger==0.9.7.1
requests==2.31.0
